requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
Django==1.8.2
requests==2.7.0
djangorestframework==3.3.0
lxml==4.2.1